server/requirements.txt
absl-py==1.3.0
aniso8601==9.0.1
astunparse==1.6.3
attrs==22.1.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
dpcpp-cpp-rt==2022.2.1
filelock==3.8.0
Fiona==1.8.22
Flask==2.1.3
flask-restx==1.0.3
flatbuffers==1.12
fonttools==4.38.0
gast==0.4.0
gdown==4.5.3
geopandas==0.10.2
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.1.0
idna==3.4
imageio==2.9.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
intel-cmplr-lic-rt==2022.2.1
intel-openmp==2022.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
keras==2.9.0rc0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
mkl==2022.2.1
mkl-random==1.2.2
munch==2.5.0
networkx==2.6.3
numpy==1.22.3
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==23.1
pandas==1.3.5
Pillow==8.0.0
pkgutil_resolve_name==1.3.10
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyproj==3.2.1
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.7.3
Shapely==1.8.5.post1
shortuuid==1.0.9
six==1.16.0
soupsieve==2.3.2.post1
tbb==2021.7.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0rc0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==1.8.0
torchaudio==0.8.0
torchvision==0.9.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
